% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_names.R
\name{change_names}
\alias{change_names}
\title{Change column names}
\usage{
change_names(data)
}
\arguments{
\item{data}{data.frame, A data frame}
}
\value{
A list of length two:

1. a data frame with the original column names and the corresponding new names

2. a data frame with the changed column names
}
\description{
Change the columns names of a data frame and stores the original column names
}
\examples{
# original data frame with 5 columns
data <- data.frame(matrix(1:20, nrow = 4, ncol = 5))
change_names(data)
}
