% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSHORT.R
\name{compareSHORT}
\alias{compareSHORT}
\title{Compare short forms obtained with different strategies}
\usage{
compareSHORT(
  a = stats::runif(10, 0.4, 2),
  b = stats::runif(10, -3, 3),
  true_theta = stats::rnorm(1000),
  theta_target = 2
)
}
\arguments{
\item{a}{vector of discrimination parameters}

\item{b}{vector of difficulty parameters}

\item{true_theta}{respondents' theta}

\item{theta_target}{number of items for the short form (it defines the theta target for the \code{cluster} and \code{guided} strategies)}
}
\value{
a list
}
\description{
Compare short forms obtained with different strategies
}
