% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_borders.R
\name{cut_borders}
\alias{cut_borders}
\title{cut_borders}
\usage{
cut_borders(x)
}
\arguments{
\item{x}{numeric/integer vector}
}
\value{
a data frame
}
\description{
Extract the limits of the intervals obtained from subsetting a vector
}
\examples{
\dontrun{
x <- seq(-3, 3, length = 5)
groups <- cut(x, 5, include.lowest = TRUE)
boundaries <- cut_borders(groups)}
}
