% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.iifs.R
\name{plot.iifs}
\alias{plot.iifs}
\title{Method for plotting the item information functions}
\usage{
\method{plot}{iifs}(x, single_panels = TRUE, items = NULL, ...)
}
\arguments{
\item{x}{\code{data.frame} of class \code{iifs} obtained with the function \code{item_info()}}

\item{single_panels}{\code{logical}, default is \code{TRUE}. Whether to show the \eqn{I_i(\theta)} of each item on a different panel}

\item{items}{default is \code{NULL} (shows all items). Allows for selecting specific items for the plot.}

\item{...}{other arguments}
}
\value{
A ggplot
}
\description{
Plot the information functions of polytomous or dichotomous items
}
\details{
If more there are more than 10 items, the legend associated to the color of the lines is not displayed.
}
\examples{
# Set random seed for reproducibility
set.seed(123)

# Simulate parameters for five dichotomous items
# according to a 2-PL specification
# b = difficulty parameters
# a = discrimination parameters
# c = lower asymptote
# e = upper asymptote
parameters <- data.frame(
  b = c(-3, -2, 0, 2, 3),
  a = runif(5, 1.2, 1.9),
  c = rep(0, 5),
  e = rep(1, 5)
)

# Compute item information functions (IIFs)
infos <- item_info(parameters)

# Plot information functions for all items
plot(infos)

# Plot information functions only for items 1 and 3
# on a single panel
plot(
  infos,
  items = c(1, 3),
  single_panels = FALSE
)
}
