% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.isa.R
\name{plot.isa}
\alias{plot.isa}
\title{Method for plotting the TIF of the test/short test form}
\usage{
\method{plot}{isa}(x, fun = "mean", show_all = FALSE, show_dist = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{isa} obtained with function \code{isa()}}

\item{fun}{\code{character}, whether to consider the mean or the sum for the computation of the TIF}

\item{show_all}{\code{logical}, default is \code{FALSE}. Whether to show the TIF of the test and the TIF target together with the TIF obtained from all the items in \code{item_par}}

\item{show_dist}{\code{logical}, default is \code{FALSE}. Whether to show or not the difference and distance (absolute difference) from the TIF target.}

\item{...}{other arguments}
}
\value{
A \code{ggplot} showing either the TIFs of the test, that of the item bank, and the TIF target or the distance/difference.
}
\description{
The test/short test form is obtained with the ISA procedure implemented with function \code{isa()}. Details on the procedure can be found in the documentation of the \code{isa()} function.
}
\examples{
# Set random seed for reproducibility
set.seed(123)

# Define the number of items in the item bank
n <- 50

# Generate latent trait values (not directly used here,
# but typically included in IRT simulations)
theta <- rnorm(500)

# Create item parameter matrix/data frame
# b = difficulty parameters
# a = discrimination parameters
# c = lower asymptote
# e = upper asymptote
item_par <- data.frame(
  b = runif(n, -3, 3),
  a = runif(n, 1.2, 1.9),
  c = rep(0, n),
  e = rep(1, n)
)

# Compute item information functions for the item bank
# and define a target test information function (TIF)
# as the mean information across items
target <- tif(item_info(item_par), fun = "mean")

# Run item selection algorithm (ISA)
# selecting a minimum of 5 items
resI <- isa(item_par, target, nmin = 5)

# Plot selected item set and related results
plot(resI)

# Show the Test Information Function (TIF)
# of the selected item bank
plot(resI, show_all = TRUE)

# Show the distance between obtained TIF and target TIF
plot(resI, show_dist = TRUE)
}
