% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.theta_target.R
\name{plot.theta_target}
\alias{plot.theta_target}
\title{Method for plotting the TIF of the test/short test form}
\usage{
\method{plot}{theta_target}(x, fun = "sum", show_targets = TRUE, ...)
}
\arguments{
\item{x}{Object of class \code{theta_target} obtained with function \code{theta_target()}}

\item{fun}{\code{character}, whether to consider the mean or the sum for the computation of the TIF}

\item{show_targets}{\code{logical}, default is \code{TRUE}. Whether to show or not the theta targets. If \code{TRUE} the theta targets are shown. The color associated to each theta target represents the specific item that has been selected for maximizing the information for that specific point.}

\item{...}{other arguments}
}
\value{
A \code{ggplot} showing the TIFs of the test, the locations of the theta targets, and the items that satisfy each theta target.
}
\description{
The test/short test form is obtained with the theta target procedure implemented with function \code{theta_target()}. Details on the procedure can be found in the documentation of the \code{theta_target()} function.
}
\details{
If more than 10 theta targets are selected, the legend associated to each theta target is not displayed.
}
\examples{
# Set random seed for reproducibility
set.seed(123)

# Define the number of items in the item bank
n <- 50

# Generate latent trait values for 500 respondents
theta <- rnorm(500)

# Create item parameter matrix/data frame
# b = difficulty parameters
# a = discrimination parameters
# c = lower asymptote
# e = upper asymptote
item_par <- data.frame(
  b = runif(n, -3, 3),
  a = runif(n, 1.2, 1.9),
  c = rep(0, n),
  e = rep(1, n)
)

# Define a theta targets
targets <- define_targets(theta, num_targets = 4)

# Perform theta-targeted item selection
resT <- theta_target(targets, item_par)

# Plot results of the target-based selection
plot(resT)

# Plot results without displaying theta target
plot(resT, show_targets = FALSE)
}
