% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_difference.R
\name{plot_difference}
\alias{plot_difference}
\title{plot_difference}
\usage{
plot_difference(difference, type = c("difference", "absolute"), n_levels = 4)
}
\arguments{
\item{difference}{data frame obtained with the function diff_theta}

\item{type}{type of difference, either as is (difference) or absolute (absolute)}

\item{n_levels}{number of levels of the starting theta (default is 4)}
}
\value{
A ggplot object
}
\description{
Plot the difference or the absolute difference between the starting theta and the theta estimated with the stf as a function of different levels of the latent trait
}
\examples{
\dontrun{
# Simulate person and item parameters
starting_theta = rnorm(100)
b <- runif(100, -3, 3)
a <- runif(100, 0.6, 2)
parameters <- data.frame(b, a)
# simulate data
data <- sirt::sim.raschtype(starting_theta, b = b, fixed.a = a)
stf <- uip(data, starting_theta = starting_theta, item_par = parameters, num_item = 5)
# compute the difference between starting theta and that estimated with the stf
my_diff <- diff_theta(stf)
# plot the difference with default number of levels
plot_difference(my_diff, type = "difference")
# plot the absolute difference with 10 levels
plot_difference(my_diff, type = "difference", n_levels = 10)
}
}
