% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTIF.R
\name{plotTIF}
\alias{plotTIF}
\title{Plot Test Information Function (TIF)}
\usage{
plotTIF(object, graph = c("all", "short", "together"))
}
\arguments{
\item{object}{Object of either class \code{shortForm} or \code{compareSHORT}}

\item{graph}{plot the TIF for the test including all items (\code{all}), the short form (\code{short}) or together (\code{together})}
}
\value{
a ggplot
}
\description{
Plot Test Information Function (TIF)
}
