% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortIRT.R
\name{shortIRT}
\alias{shortIRT}
\title{Select item for the short forms using observed data}
\usage{
shortIRT(
  data,
  num_item = NULL,
  strategy = c("smart", "cluster", "guided"),
  fixed.b = NULL,
  fixed.a = NULL,
  true_theta = NULL
)
}
\arguments{
\item{data}{data frame with observed responses}

\item{num_item}{number of desired items for the short version}

\item{strategy}{which strategy for selecting the items?}

\item{fixed.b}{vector of fixed difficulty parameters}

\item{fixed.a}{vector of fixed discrimination parameters}

\item{true_theta}{vector of respondent parameters}
}
\value{
a data set
}
\description{
Select item for the short forms using observed data
}
