% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.isa.R
\name{summary.isa}
\alias{summary.isa}
\title{Method for the summary of the test/short test form}
\usage{
\method{summary}{isa}(object, ...)
}
\arguments{
\item{object}{Object of class \code{isa}}

\item{...}{other arguments}
}
\value{
A summary of the test obtained from the application of ISA
}
\description{
The test/short test form is obtained with the ISA procedure implemented with function \code{isa()}. Details on the procedure can be found in the documentation of the \code{isa()} function.
}
\examples{
# Set random seed for reproducibility
set.seed(123)

# Define the number of items in the item bank
n <- 50

# Create item parameter matrix/data frame
# b = difficulty parameters
# a = discrimination parameters
# c = lower asymptote
# e = upper asymptote
item_par <- data.frame(
  b = runif(n, -3, 3),
  a = runif(n, 1.2, 1.9),
  c = rep(0, n),
  e = rep(1, n)
)

# Compute item information functions and define
# a target Test Information Function (TIF)
# using the mean information across items
target <- tif(
  item_info(item_par),
  fun = "mean"
)

# Run item selection algorithm (ISA)
# selecting at least 5 items
resI <- isa(
  item_par,
  target,
  nmin = 5
)

# Summarize item selection results
summary(resI)
}
