% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoryIRT.R
\name{theoryIRT}
\alias{theoryIRT}
\title{Select items for the short forms using only item and respondent parameters}
\usage{
theoryIRT(
  a = stats::runif(10, 0.4, 2),
  b = stats::runif(10, -3, 3),
  strategy = c("smart", "cluster", "guided"),
  true_theta = stats::rnorm(1000),
  theta_target = 2
)
}
\arguments{
\item{a}{vector of discrimination parameters}

\item{b}{vector of difficulty parameters}

\item{strategy}{select the strategy for selecting the items, \code{smart}, \code{cluster}, \code{guided}}

\item{true_theta}{respondents' theta}

\item{theta_target}{number of items for the short form (it defines the theta target for the \code{cluster} and \code{guided} strategies)}
}
\value{
a list
}
\description{
Select items for the short forms using only item and respondent parameters
}
