% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench.R
\name{bench}
\alias{bench}
\title{Benchmark Procedure}
\usage{
bench(item_par = NULL, iifs = NULL, theta = NULL, num_item = NULL)
}
\arguments{
\item{item_par}{data.frame, dataframe with nrows equal to the number of items and 4 columns, one for each of the item parameters. The columns must be named "a", "b", "c", "e" and must contain the respective parameters.}

\item{iifs}{data.frame, dataframe with nrows equal to the lenght of theta and ncols equal to the number of items, it contains the IIFs of the items in the full-length test. Cannot use both ipar and iifs}

\item{theta}{numeric, vector with the latent trait values}

\item{num_item}{integer, the number of items to include in the short test form}
}
\value{
A list of length 3 with:

- stf: dataframe with the items selected for inclusion in the STF (column isel), their maximum information function (maxiif), for  a specific theta (column theta)
- item_par: the original dataframe containing the item parameters
- selected_items: dataframe with the parameters of the selected items
}
\description{
Create a Short Test Form (STF) using a benchmark procedure (i.e., the n most informative items are selected, where n is the number of items to include in the STF)
}
\examples{
set.seed(123)
n <- 50
theta <- rnorm(500, sd = 2)
item_par <- data.frame(
  b = runif(n, -3, 3),
  a = runif(n, 1.2, 1.9),
  c = rep(0, n),
  e = rep(1, n)
)
resB <- bench(item_par, theta = theta, num_item = 5)
str(resB)
}
