% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_borders.R
\name{cut_borders}
\alias{cut_borders}
\title{Cut borders}
\usage{
cut_borders(x)
}
\arguments{
\item{x}{numeric/integer vector}
}
\value{
A data frame with two columns. The first column contains the lower bounds of each interval. The second column contains the upper bound of each interval
}
\description{
Extract the limits of the intervals obtained from sub setting a vector
}
\examples{
x <- seq(-3, 3, length = 5)
groups <- cut(x, 5, include.lowest = TRUE)
boundaries <- cut_borders(groups)
}
