% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_targets.R
\name{define_targets}
\alias{define_targets}
\title{Define theta targets}
\usage{
define_targets(theta, num_targets = NULL, method = c("equal", "clusters"))
}
\arguments{
\item{theta}{numeric vector defining the latent trait theta}

\item{num_targets}{integer, define the number of theta targets to obtain}

\item{method}{character, either equal (default) or clusters}
}
\value{
A vector of length num_targets with the generated targets
}
\description{
Define theta targets according to two procedures.
}
\examples{
set.seed(123)
theta <- rnorm(1000)
targets <- define_targets(theta, num_targets = 5, method = "clusters")
}
