% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{eip}
\alias{eip}
\title{eip}
\usage{
eip(
  data,
  fixed.b = NULL,
  fixed.a = NULL,
  seed = 999,
  true_theta = NULL,
  num_item = NULL
)
}
\arguments{
\item{data}{a subject x item matrix}

\item{fixed.b}{a vector of fixed difficulty parameters. If empty, the item parameters estimated from the data will be used.}

\item{fixed.a}{a vector of fixed discrimination parameters. If empty, the item parameters estimated from the data will be used.}

\item{seed}{a random seed (default is 999)}

\item{true_theta}{a vector with length equal to the number of rows in data with the true theta values of the subjects. If empty, the theta values will be estimated from the data.}

\item{num_item}{the number of item to included in the short test form}
}
\value{
A list of length 3
}
\description{
Create a Short Test Form (STF) using the theta-target procedure based on the equal segmentation of the latent trait
}
