% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{eip}
\alias{eip}
\title{eip}
\usage{
eip(
  data,
  item_par = NULL,
  seed = 999,
  starting_theta = NULL,
  num_item = NULL,
  theta_targets = NULL
)
}
\arguments{
\item{data}{a subject x item matrix}

\item{item_par}{Two column matrix with the item parameters. The first column must contain the difficulty parameters, the second column must contain the discrimination parameters.}

\item{seed}{a random seed (default is 999)}

\item{starting_theta}{a vector with length equal to the number of rows in data with the true theta values of the subjects. If empty, the theta values will be estimated from the data.}

\item{num_item}{the number of item to included in the short test form}

\item{theta_targets}{A vector of length n > 2 (where n is the number of item to be included in the short test form) with specific theta targets. Might also be the same theta target repeated for as many time as the number of item to be included in the short test form}
}
\value{
A list of length 5:
         - item_stf: data frame with nrows equal to the number of items included in the STF, contains the items, the theta targets and the IIF in respect to the theta target
         - summary: data frame with two rows, contains the list of items included in the STF and the total information, along with the information of the full length test
         - info_stf: contains the information for each level of the latent trait for the STF
         - info_full: contains the information for each level of the latent trait for the full-length test
         - theta: data frame with the starting theta and the theta estimated with the STF
}
\description{
Create a Short Test Form (STF) using the theta-target procedure based on the equal segmentation of the latent trait (Equal Interval Procedure, EIP)
}
\examples{
\dontrun{
# Simulate person and item parameters
starting_theta = rnorm(100)
b <- runif(100, -3, 3)
a <- runif(100, 0.6, 2)
parameters <- data.frame(b, a)
# simulate data
data <- sirt::sim.raschtype(starting_theta, b = b, fixed.a = a)
stf <- eip(data, starting_theta = starting_theta,
item_par = parameters, num_item = 5)
# check the obtained short test form
stf$item_stf
# check the comparison between the short test form and the full-length test
stf$summary

# Short test form with cut off values
stf_cutoff <- eip(data, starting_theta = starting_theta,
item_par = parameters, theta_targets = rep(2, 5))
stf_cutoff$item_stf
}
}
