% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_info.R
\name{i_info}
\alias{i_info}
\title{Item Information Function (single item, IIF)}
\usage{
i_info(item_pars, theta = seq(-5, 5, length.out = 1000), K = NULL)
}
\arguments{
\item{item_pars}{Numeric matrix with one row corresponding to one item.
 For dichotomous items, the matrix must have 4 columns, one for each of the item parameters. The columns must be named "a", "b", "c", "e" and must contain the respective IRT parameters, namely discrimination \eqn{a_i}, location \eqn{b_i}, pseudo-guessing \eqn{c_i}, and upper asymptote \eqn{e_i}.
 For polytomous items, the matrix has \eqn{2K} columns. The first \eqn{K} columns correspond to step
discrimination parameters \eqn{a_1, \dots, a_K} (must be named "a"), and the last \eqn{K}
columns correspond to step difficulty (threshold) parameters (must be named "b")
\eqn{b_1, \dots, b_K}.}

\item{theta}{Numeric vector of latent trait values. Default is a vector of a thousand values ranging from -5 to +5}

\item{K}{Integer. Number of thresholds for  the categories of the polytoumous items (i.e., number of categories minus one). Default is \code{NULL} (assumes dichotomous items).}
}
\value{
A numeric vector of length equal to \code{theta}, which contains the item information function for a single item with respect to the values specified in \code{theta}
}
\description{
Compute the item information function (IIF) for a single dichotomous
or polytoumous item under either the 4-PL model (dichotomous item) or  a Generalized Partial Credit (polytomous item).
Specific models (e.g., 3-PL, 2-PL, 1-PL, or PCM, Rating Scale) are obtained by imposing
constraints on the item parameters.
See \code{Details}.
}
\details{
Let \eqn{P(\theta)} denote the  probability of a correct response \eqn{x_{pi} = 1} for person \eqn{p} on item \eqn{i} under the four-parameter logistic
(4-PL; Barton & Lord, 1981) model is defined as:

\deqn{
P(\theta) =
c_i + \frac{e_i - c_i}{1 + \exp\left[-a_i(\theta_p - b_i)\right]}
}

where \eqn{a_i} is the discrimination parameter,
\eqn{b_i} is the difficulty parameter (or location of item \eqn{i} on the latent trait),
\eqn{c_i} is the lower asymptote (pseudo-guessing probability),
and \eqn{e_i} is the upper asymptote (inattention/slip). By constraining \eqn{e_i = 1}, \eqn{c_i = 0}, and \eqn{a_i=1} \eqn{\forall i}, the probability
is computed according to the 3-PL (Lord, 1980), 2-PL (Birnbaum, 1968) and 1-PL, respectively.

Let \eqn{Q(\theta) = 1 - P(\theta)}, the information function of item \eqn{i} is computed as:

\deqn{
I_i(\theta) =
\frac{a_i^2 \left[P(\theta) - c_i\right]^2 \left[e_i - P(\theta)\right]^2}
{(e_i - c_i)^2 \, P(\theta) \, Q(\theta)}
}

According to the Generalized Partial Credit Model (GPCM; Muraki, 1997), for a polytomous item with \eqn{K} thresholds separating the \eqn{K + 1} categories, the probability of category \eqn{k} is defined as:
\deqn{
P(Y = k \mid \theta) =
\frac{\exp\left( \sum_{k=1}^K a_k (\theta - b_k) \right)}
{\sum_{j=0}^K \exp\left( \sum_{k=1}^K a_k (\theta - b_k) \right)}
}

where \eqn{a_k} and \eqn{b_k} are the discrimination and location parameters associate with each threshold \eqn{k}. If \eqn{a_k = 1, \, \forall k}, the Partial Credit Model (PCM, Muraki, 1992) is obtained.

The item information is computed as:
\deqn{
I_i(\theta) = \sum_{k=0}^K \frac{[P'_k(\theta)]^2}{P_k(\theta)}
}
}
\examples{

theta <- seq(-4, 4, length.out = 200)

# IIF of an item with b = 0
item_par <- data.frame(b = 0, a = 1.5, c = .10, e = .98)
info_dichotomous <- i_info(item_par, theta = theta)

# 4-category item (K = 3)
item_pars <- data.frame(a1 = 1.2, a2 = 1.0, a3= 0.8,
                        b1 = -1.0, b2 = 0.0, b3 = 1.2)

info <- i_info(item_pars, theta = theta, K = 3)
}
\references{
Barton, M. A., & Lord, F. M. (1981). An upper asymptote for the
three-parameter logistic item-response model. ETS Research Report
Series, 1981(1), i–8. Princeton, NJ: Educational Testing Service.

Birnbaum, A. (1968). Some latent trait models and their use in inferring
an examinee's ability. In F. M. Lord & M. R. Novick (Eds.),
Statistical theories of mental test scores (pp. 397–479).
Reading, MA: Addison-Wesley.

Lord, F. M. (1980). Applications of item response theory to practical
testing problems. Hillsdale, NJ: Lawrence Erlbaum Associates.

Muraki, G. (1992). A generalized partial credit model: Application of an EM algorithm.
Psychometrika, 57(2), 159–176.

Muraki, G. (1997). A generalized partial credit model with step discrimination.
Journal of Educational Measurement, 34(2), 115–127.
}
