% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_info.R
\name{i_info}
\alias{i_info}
\title{Itenm information function (single item)}
\usage{
i_info(b, a = 1, c = 0, e = 1, theta = seq(-5, 5, length.out = 1000))
}
\arguments{
\item{b}{location of item i, with no default.}

\item{a}{discrimination parameter for item i. Default is 1.}

\item{c}{pesudoguessing parameter of item i. Default is 0.}

\item{e}{Upper asymptote of item i. Default is 1.}

\item{theta}{numeric vector (might also be a single value) representing the latent trait. Default is a vector of 1 thousand values rangin from -5 to +5}
}
\value{
A numeric vector of length equal to theta, which contains the IIF for a single item with respect to the values specified in theta
}
\description{
Compute the item information function for a single item. The formula for the computation is taken from REFRENCE d magis
}
\examples{
# IIF of an item with b = 0
i_info(b = 0, theta = c(-3,-1,0,1,3))
}
