% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT.R
\name{IRT}
\alias{IRT}
\title{Compute expected probability for a single dichotomous item}
\usage{
IRT(theta, b = 0, a = 1, c = 0, e = 1)
}
\arguments{
\item{theta}{\code{numeric} latent trait level of person \eqn{p}. It can be a single value or a vector of values.}

\item{b}{\code{numeric} location of item \eqn{i}. Default is 0.}

\item{a}{\code{numeric} discrimination parameter for item \eqn{i}. Default is 1.}

\item{c}{\code{numeric} pesudoguessing parameter of item \eqn{i}. Default is 0.}

\item{e}{\code{numeric} upper asymptote of item \eqn{i}. Default is 1.}
}
\value{
a single value, that is the probability of the correct response for item \eqn{i} given the specified parameters
}
\description{
Compute the expected probability for an item \eqn{i} according to its IRT parameters.
Depending on the parameters that are specified, the probability is computed according to the 1-PL, 2-PL, 3-PL, or 4-PL models.
}
\details{
The probability of a correct response \eqn{x_{pi} = 1} for person \eqn{p} on item \eqn{i} under the four-parameter logistic
(4-PL; Barton & Lord, 1981) model is defined as:

\deqn{
P(x_{pi} = 1 \mid \theta_p, b_i, a_i, c_i, e_i) =
c_i + \frac{e_i - c_i}{1 + \exp\left[-a_i(\theta_p - b_i)\right]}
}

where \eqn{a_i} is the discrimination parameter,
\eqn{b_i} is the difficulty parameter (or location of item \eqn{i} on the latent trait),
\eqn{c_i} is the lower asymptote (pseudo-guessing probability),
and \eqn{e_i} is the upper asymptote (inattention/slip). By constraining \eqn{e_i = 1}, \eqn{c_i = 0}, and \eqn{a_i=1} \eqn{\forall i}, the probability
is computed according to the 3-PL (Lord, 1980), 2-PL (Birnbaum, 1968) and 1-PL, respectively.
}
\examples{
# compute the probability for an item according to 1-PL model
IRT(theta = 0, b = 0,  a = 1, c = 0, e = 1)
# compute probability for a vector of thetas for the same item
IRT(theta = c(-1, 0, 1), b = 0,  a = 1, c = 0, e = 1)
# compute probability for a vector of thetas for an item according to the 4-PL model
IRT(theta = c(-1, 0, 1), b = 0,  a = 1.25, c = 0.10, e = 0.98)
}
\references{
Barton, M. A., & Lord, F. M. (1981). An upper asymptote for the
three-parameter logistic item-response model. ETS Research Report
Series, 1981(1), i–8. Princeton, NJ: Educational Testing Service.

Birnbaum, A. (1968). Some latent trait models and their use in inferring
an examinee's ability. In F. M. Lord & M. R. Novick (Eds.),
Statistical theories of mental test scores (pp. 397–479).
Reading, MA: Addison-Wesley.

Lord, F. M. (1980). Applications of item response theory to practical
testing problems. Hillsdale, NJ: Lawrence Erlbaum Associates.
}
