% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT.R
\name{IRT}
\alias{IRT}
\title{Compute expected probability}
\usage{
IRT(theta, b = 0, a = 1, c = 0, e = 1)
}
\arguments{
\item{theta}{latent trait level of person p, Can be a single value or a vector.}

\item{b}{location of item i. Default is 0.}

\item{a}{discrimination parameter for item i. Default is 1.}

\item{c}{pesudoguessing parameter of item i. Default is 0.}

\item{e}{Upper asymptote of item i. Default is 1.}
}
\value{
a single value, that is the probability of the correct response for item i given the specified parameters
}
\description{
Compute the expected probability for an item i according to its IRT parameters.
According to the parameters that are specified, the probability according to the 1-PL, 2-PL, 3-PL, or 4-PL models is computed
}
\examples{
IRT(theta = 0, b = 0,  a = 1, c = 0, e = 1)
}
