% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isa.R
\name{isa}
\alias{isa}
\title{Item Selection Algorithm}
\usage{
isa(item_pars, tif_target, nmin = round(nrow(item_pars) * 0.1), K = NULL)
}
\arguments{
\item{item_pars}{\code{data.frame}, dataframe with number of rows equal to the number of items.
 For dichotomous items, the dataframe must have 4 columns, one for each of the item parameters. The columns must be named "a", "b", "c", "e" and must contain the respective IRT parameters, namely discrimination \eqn{a_i}, location \eqn{b_i}, pseudo-guessing \eqn{c_i}, and upper asymptote \eqn{e_i}.
 For polytomous items, the dataframe has \eqn{2K} columns, where \eqn{K} is the number of thresholds of the items (number of response categorie \eqn{- 1}). The first \eqn{K} columns correspond to step
discrimination parameters \eqn{a_1, \dots, a_K} (must be named "a"), and the last \eqn{K}
columns correspond to step difficulty (threshold) parameters
\eqn{b_1, \dots, b_K} (must be named "b").}

\item{tif_target}{\code{data.frame} with two columns: \code{theta} the latent trait \eqn{\theta} and \code{tif} defining the values of the tif target. The TIF target should be computed as the mean TIF to allow for the comparability with the TIF obtained from the test.}

\item{nmin}{\code{numeric}, minimum number of items to be included in the test (i.e., the termination criterion is not tested until the minimum number of items is reached). Default is the 10\% of the total number of items.}

\item{K}{\code{integer}, number of thresholds for polytomous items (number of response categories minus 1). Default is \code{NULL} (assumes dichotomous items).}
}
\value{
An object of class \code{isa} of length 6 containing:

\itemize{
  \item \strong{test}: a dataframe containing the items selected for inclusion
  in the test (column \code{isel}) and the minimum number of items set in \code{nmin}.

  \item \strong{item_pars}: the original dataframe containing the item parameters.

  \item \strong{selected_items}: a dataframe containing the parameters of the selected items.

  \item \strong{all_iifs}: a dataframe containing the IIFs of all the original items.

  \item \strong{tif_target}: the dataframe with the TIF target used for the item selection.

  \item \strong{K}: Number of thresholds for the response categories of the items. If the items are dichotomous \code{K} is \code{NULL}.
}
}
\description{
Develop a test or a short form given the parameters of dichotomous or polytomous in an item bank/full-length test according to the Item Selection procedure (ISA, Epifania & Finos, 2025). See \code{Details}.
}
\details{
Let \eqn{t = 0, \dots, T} denote the iteration index of the procedure, \eqn{\text{TIF}^*} denote the test information target, and \eqn{\text{TIF}^t} denote the test information function obtained from \eqn{Q^t \subset B} (where \eqn{B} is the item bank and \eqn{Q^t} is the subset of items selected up to iteration \eqn{t}).
At \eqn{t = 0}: \eqn{\text{TIF}^0(\theta) = 0}, \eqn{\forall \theta}, \eqn{Q^0 = \emptyset}.
For \eqn{t \geq 0},

\enumerate{
  \item Consider the available items at iteration \eqn{t}

  \deqn{A^t = B \setminus Q^t}

  \item Compute the provisional TIF (\eqn{\text{pTIF}_i}) for each of the available items

  \deqn{\forall i \in A^t, \text{pTIF}_{i} := \frac{\text{TIF}^t + I_{i}(\theta)}{|Q^t|+1}}

  \item Select a provisional item \eqn{i^*} allowing for minimizing the distance from the TIF target

  \deqn{i^* := \arg \min_{i \in A^t} \text{abs}(\text{TIF}^* - \text{pTIF}_i)}

  \item Test the termination criterion: If

  \eqn{\text{abs}(\text{TIF}^* - \text{pTIF}_{i^*}) \ngeq \text{abs}(\text{TIF}^* - \text{TIF}^{t}), Q^{t+1} = Q^{t} \cup \{i^+\}}, \eqn{\text{TIF}^{t+1} = \text{pTIF}_{i^+}}, Go back to 1

  \item Stop, \eqn{Q_{\text{ISA}} = Q^t}

}

Further details on the algorithm can be found in Epifania & Finos (2025), where the algorithm is denoted as Frank.
}
\examples{
set.seed(123)

n <- 50
theta <- rnorm(500)

item_pars <- data.frame(
  b = runif(n, -3, 3),
  a = runif(n, 1.2, 1.9),
  c = rep(0, n),
  e = rep(1, n)
)

tif_target <- tif(
  item_info(item_pars[-c(3, 5, 10, 14), ]),
  fun = "mean"
)

resIsa <- isa(item_pars, tif_target, nmin = 4)
str(resIsa)

# Polytomous items with user-defined theta targets

item_pars <- data.frame(
  matrix(c(
    1.2, 1.0, 0.8,  -1.0,  0.0,  1.2,
    0.9, 1.1, 1.3,  -0.5,  0.7,  1.8,
    0.5, 1.5, 1.0,  -1.5, -1.0,  0.0,
    1.0, 1.0, 1.0,  -1.5,  0.0,  0.5
  ), nrow = 4, byrow = TRUE)
)

colnames(item_pars) <- paste(
  rep(c("a", "b"), each = 3),
  1:3,
  sep = ""
)

resT_poly <- theta_target(c(-1, 0), item_pars, K = 3)
str(resT_poly)
}
\references{
Epifania, O. M., & Finos, L. (2025). Nothing lasts forever – only item
administration: An item response theory algorithm to shorten tests.
In E. Di Bella, V. Gioia, C. Lagazio, & S. Zaccarin (Eds.),
Statistics for Innovation III (pp. 188–193).
Italian Statistical Society Series on Advances in Statistics.
Springer, Cham. https://doi.org/10.1007/978-3-031-95995-0_32
}
