% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isa.R
\name{isa}
\alias{isa}
\title{Item Selection Algorithm}
\usage{
isa(item_pars, tif_target, nmin = round(nrow(item_pars) * 0.1), K = NULL)
}
\arguments{
\item{item_pars}{\code{data.frame}, dataframe with nrows equal to the number of items. #'
 For dichotomous items, the matrix must have 4 columns, one for each of the item parameters. The columns must be named "a", "b", "c", "e" and must contain the respective IRT parameters, namely discrimination \eqn{a_i}, location \eqn{b_i}, pseudo-guessing \eqn{c_i}, and upper asymptote \eqn{e_i}.
 For polytomous items, the matrix has \eqn{2K} columns. The first \eqn{K} columns correspond to step
discrimination parameters \eqn{a_1, \dots, a_K} (must be named "a"), and the last \eqn{K}
columns correspond to step difficulty (threshold) parameters (must be named "b")
\eqn{b_1, \dots, b_K}.}

\item{tif_target}{\code{data.frame} with two columns: \code{theta} the latent trait \eqn{\theta} and \code{tif} defining the values of the tif target. The TIF target should be computed as the mean TIF to allow for the comparability with the TIF obtained from the STF.}

\item{nmin}{\code{numeric}, minimum number of items to be included in the STF (i.e., the termination criterion is not tested until the minimum number of items is reached). Default is the 10% of the total number of items.}

\item{K}{Integer. Number of thresholds for  the categories of the polytoumous items (i.e., number of categories minus one). Default is \code{NULL} (assumes dichotomous items).}
}
\value{
An object of class \code{isa} of length 4 containing:

\itemize{
  \item \strong{stf}: a data frame containing the items selected for inclusion
  in the short test form (column \code{isel}) and the minimum number of items set in \code{nmin}.

  \item \strong{item_pars}: the original data frame containing the item
  parameters.

  \item \strong{all_iifs}: a data frame containing the IIFs of all the original items

  \item \strong{tif_target}:the data frame with the TIF traget used for teh item selection

  \item{\code{K}: Number of thresholds for the response categories of the items. If the items are dichotomous \code{K} is \code{NULL}.}
}
}
\description{
Develop a short test form given the item parameters (dichotomous or polytomous) according to the Item Selection procedure. See \code{Details}.
}
\examples{
set.seed(123)
n <- 50
theta <- rnorm(500)
item_pars <- data.frame(
  b = runif(n, -3, 3),
  a = runif(n, 1.2, 1.9),
  c = rep(0, n),
  e = rep(1, n)
)
tif_target <- tif(item_info(item_pars[-c(3,5,10,14), ]), fun = "mean")
resIsa <- isa(item_pars, tif_target, nmin = 4)
str(resIsa)
# polytomous items with user defined theta targets
item_pars <- data.frame(matrix(c(
 1.2, 1.0, 0.8,  -1.0, 0.0, 1.2,
 0.9, 1.1, 1.3,  -0.5, 0.7, 1.8,
 0.5, 1.5, 1, -1.5, -1.0, 0,
 1, 1, 1, -1.5, -0, 0.5
 ), nrow = 4, byrow = TRUE))
colnames(item_pars) = paste(rep(c("a", "b"), each = 3), 1:3, sep = "")
resT_poly <- theta_target(c(-1,0), item_pars, K = 3)
str(resT_poly)
}
