% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_info.R
\name{item_info}
\alias{item_info}
\title{Item Information Functions (multiple items, IIFs)}
\usage{
item_info(item_pars, theta = seq(-5, 5, length.out = 1000), K = NULL)
}
\arguments{
\item{item_pars}{\code{data.frame}, dataframe with nrows equal to the number of items. #'
 For dichotomous items, the matrix must have 4 columns, one for each of the item parameters. The columns must be named "a", "b", "c", "e" and must contain the respective IRT parameters, namely discrimination \eqn{a_i}, location \eqn{b_i}, pseudo-guessing \eqn{c_i}, and upper asymptote \eqn{e_i}.
 For polytomous items, the matrix has \eqn{2K} columns. The first \eqn{K} columns correspond to step
discrimination parameters \eqn{a_1, \dots, a_K} (must be named "a"), and the last \eqn{K}
columns correspond to step difficulty (threshold) parameters (must be named "b")
\eqn{b_1, \dots, b_K}.}

\item{theta}{Numeric vector of latent trait values. Default is a vector of A thousand values ranging from -5 to +5}

\item{K}{Integer. Number of thresholds for  the categories of the polytoumous items (i.e., number of categories minus one). Default is \code{NULL} (assumes dichotomous items).}
}
\value{
A \code{matrix} of class \code{iifs} with nrows equal to the length of \code{theta} and ncols equal to the number of items in \code{item_par}
}
\description{
Computes the item information functions for multiple dichotous or polytomous items
}
\details{
Let \eqn{P(\theta)} denote the probability of a correct response under the
four-parameter logistic (4PL) model (dichotomous responses):

\deqn{
P(\theta) =
c_i + \frac{e_i - c_i}{1 + \exp\left[-a_i(\theta - b_i)\right]}
}

and let \eqn{Q(\theta) = 1 - P(\theta)}.

The information function of item \eqn{i} is computed as:

\deqn{
I_i(\theta) =
\frac{a_i^2 \left[P(\theta) - c_i\right]^2 \left[e_i - P(\theta)\right]^2}
{(e_i - c_i)^2 \, P(\theta) \, Q(\theta)}
}

For a polytomous item with \eqn{K} thresholds separating the \eqn{K + 1} categories, the probability of category \eqn{k} is defined as:
\deqn{
P(Y = k \mid \theta) =
\frac{\exp\left( \sum_{s=1}^k a_i (\theta - b_i) \right)}
{\sum_{j=0}^K \exp\left( \sum_{s=1}^j a_i (\theta - b_i) \right)}
}

with the convention that the empty sum (for \eqn{k = 0}) is equal to zero.

The item information is computed as:
\deqn{
I_i(\theta) = \sum_{k=0}^K \frac{[P'_k(\theta)]^2}{P_k(\theta)}
}
}
\examples{
set.seed(123)
parameters <- data.frame(b = c(-3,-2,0, 2, 3),
a = runif(5, 1.2, 1.9),
c = rep(0,5),
e= rep(1, 5))
# compute information for dichomtous items with default theta values
infos <- item_info(parameters)
head(infos)
# 4 items with 4 response categories (K = 3)
item_pars <- data.frame(matrix(c(
        1.2, 1.0, 0.8,  -1.0, 0.0, 1.2,
        0.9, 1.1, 1.3,  -0.5, 0.7, 1.8,
       0.5, 1.5, 1, -1.5, -1.0, 0,
       1, 1, 1, -1.5, -0, 0.5),
       nrow = 4,
       byrow = TRUE))
colnames(item_pars) <- paste(rep(c("a", "b"), each = 3), 1:3, sep = "")
info_poly <- item_info(item_pars, K = 3)
head(info_poly)
}
