% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_info.R
\name{item_info}
\alias{item_info}
\title{Item Information function (multiple items)}
\usage{
item_info(ipar, theta = seq(-5, 5, length.out = 1000))
}
\arguments{
\item{ipar}{dataframe with nrows equal to the number of items and 4 columns, one for each of the item parameters. The columns must be named "a", "b", "c", "e" and must contain the respective parameters.}

\item{theta}{numeric vector (might also be a single value) representing the latent trait. Default is a vector of 1 thousand values ranging from -5 to +5}
}
\value{
A matrix with nrows equal to the length of theta and ncols equal to the number of items in ipar.
}
\description{
Item Information function (multiple items)
}
\examples{
set.seed(123)
parameters <- data.frame(b = c(-3,-2,0, 2, 3),
a = runif(5, 1.2, 1.9),
c = rep(0,5),
e= rep(1, 5))
infos <- item_info(parameters)
head(infos)
}
