% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bench.R
\name{plot.bench}
\alias{plot.bench}
\title{Method for plotting the TIF of the test/short test form}
\usage{
\method{plot}{bench}(x, fun = "sum", ...)
}
\arguments{
\item{x}{Object of class \code{bench}}

\item{fun}{\code{character}, whether to consider the mean or the sum for the computation of the TIF}

\item{...}{other arguments}
}
\value{
A \code{ggplot} showing the TIFs of the test.
}
\description{
The test/short test form is obtained with the benchmark procedure implemented with function \code{bench()}. Details on the procedure can be found in the documentation of the \code{bench()} function.
}
\examples{
set.seed(123)
n <- 50
theta <- rnorm(500, sd = 2)
item_par <- data.frame(
  b = runif(n, -3, 3),
  a = runif(n, 1.2, 1.9),
  c = rep(0, n),
  e = rep(1, n)
)
resB <- bench(item_par, theta = theta, num_item = 5)
plot(resB)
# plot only the TIF of the test
plot(resB, show_both = FALSE)
# same but with polytomous items
item_pars <- data.frame(matrix(c(
 1.2, 1.0, 0.8,  -1.0, 0.0, 1.2,
 0.9, 1.1, 1.3,  -0.5, 0.7, 1.8,
 0.5, 1.5, 1, -1.5, -1.0, 0,
 1, 1, 1, -1.5, -0, 0.5
 ), nrow = 4, byrow = TRUE))
colnames(item_pars) = paste(rep(c("a", "b"), each = 3), 1:3, sep = "")
resB_poly <- bench(item_pars, theta = theta, num_item = 2, K = 3)
plot(resB_poly)
}
