% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.isa.R
\name{plot.isa}
\alias{plot.isa}
\title{Method for plotting the TIF of the STF}
\usage{
\method{plot}{isa}(x, fun = "mean", show_all = TRUE, show_dist = FALSE, K = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{theta_target} obtained with function \code{theta_target()}}

\item{fun}{\code{character}, whether to consider the mean or the sum for the computation of the TIF}

\item{show_all}{\code{logical}, default is \code{TRUE}. Whether to show or not the TIF obtained from the full-length test}

\item{show_dist}{\code{logical}, default is TRUE. Whether to show or not the theta targets}

\item{...}{other arguments}

\item{theta}{\code{numeric}, latent trait for the graphical representation}
}
\value{
A \code{ggplot} showing the TIFs of both the STF and the full-length test
}
\description{
The STF is obtained with the ISA procedure
}
\examples{
set.seed(123)
n <- 50
theta <- rnorm(500)
item_par <- data.frame(
  b = runif(n, -3, 3),
  a = runif(n, 1.2, 1.9),
  c = rep(0, n),
  e = rep(1, n)
)
targets <- define_targets(theta, num_targets = 4)
resT <- theta_target(targets, item_par)
plot(resT)
# plot without showing the theta targets
plot(resT, show_targets = FALSE)
}
