% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tif.R
\name{plot_tif}
\alias{plot_tif}
\title{plot_tif}
\usage{
plot_tif(results, tif = c("stf", "full", "both"))
}
\arguments{
\item{results}{A list of length 4 obtained with the functions for shortening tests}

\item{tif}{character to indicate the TIF to plot, either stf (default), full test or both}
}
\value{
A ggplot object
}
\description{
Plot the test information function of the short test form (default), of the full length test or of both versions
}
\examples{
\dontrun{
# Simulate person and item parameters
true_theta <- rnorm(100)
b <- runif(100, -3, 3)
a <- runif(100, 0.6, 2)
parameters <- data.frame(b, a)
# simulate data
data <- sirt::sim.raschtype(true_theta, b = b, fixed.a = a)
stf <- uip(data, true_theta = true_theta, item_par = parameters, num_item = 5)
# plot the test information function of the full-length test
plot(stf, tif = "full")
# plot the test information of the full-length test and of the short test form
plot(stf, tif = "both")
}
}
