% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.isa.R
\name{summary.isa}
\alias{summary.isa}
\title{Method for the summary of the STF}
\usage{
\method{summary}{isa}(object, ...)
}
\arguments{
\item{object}{Object of class \code{isa}}

\item{...}{other arguments}
}
\value{
A summary of the STF obtained from the application of the ISA procedure
}
\description{
The STF is obtained with the ISA procedure implemented with function \code{isa()}.Details on the procedure can be found in the documentation of the \code{isa()} function.
}
\examples{
set.seed(123)
n <- 50
theta <- rnorm(500)
item_par <- data.frame(
  b = runif(n, -3, 3),
  a = runif(n, 1.2, 1.9),
  c = rep(0, n),
  e = rep(1, n)
)
target <- tif(item_info(item_par), fun = "mean")
resI <- isa(item_par, target, nmin = 5)
summary(resI)
}
