% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_target.R
\name{theta_target}
\alias{theta_target}
\title{Theta target procedure}
\usage{
theta_target(
  targets,
  item_pars,
  theta = seq(-5, 5, length.out = 1000),
  K = NULL
)
}
\arguments{
\item{targets}{\code{numeric}, either a vector with the discrete values of theta for which the information needs to be
maximized obtained with the \code{define_targets()} function or a vector with user-defined values.
If the same theta value is defined and repeated several time, it can be passed as a named list, where \code{value} indicates the value
that needs to be repeated and \code{num_targets} defines the number of times it is repeated for.}

\item{item_pars}{\code{data.frame} with nrows equal to the number of items.
 For dichotomous items, the matrix must have 4 columns, one for each of the item parameters. The columns must be named "a", "b", "c", "e" and must contain the respective IRT parameters, namely discrimination \eqn{a_i}, location \eqn{b_i}, pseudo-guessing \eqn{c_i}, and upper asymptote \eqn{e_i}.
 For polytomous items, the matrix has \eqn{2K} columns, where \eqn{K} is the number of thresholds of the items (number of response categorie \eqn{- 1}). The first \eqn{K} columns correspond to step
discrimination parameters \eqn{a_1, \dots, a_K} (must be named "a"), and the last \eqn{K}
columns correspond to step difficulty (threshold) parameters (must be named "b")
\eqn{b_1, \dots, b_K}.}

\item{theta}{\code{numeric} vector with the values of the latent trait \eqn{\theta} (needed for the computation of the IIFs of all items)}

\item{K}{\code{integer} defining the number of thresholds for  the categories of the polytoumous items (i.e., number of categories minus one). Default is \code{NULL} (assumes dichotomous items).}
}
\value{
An object of class \code{theta_target} of length 4 containing:

\itemize{
  \item \strong{test}: a data frame containing the items selected for inclusion
  in the test (column \code{isel}), their maximum item information
  function (column \code{maxiif}), and the corresponding theta target
  (column \code{theta_target}).

  \item \strong{item_pars}: the original data frame containing the item
  parameters.

  \item \strong{selected_items}: a data frame containing the parameters of the
  selected items.

  \item \strong{intervals}: a character string indicating how the theta
  targets were obtained. The value \code{"clusters"} denotes clustering of
  the latent trait, \code{"equal"} denotes equally spaced intervals, and
  \code{"unknown"} identifies any other case (e.g., user-defined theta targets).

  \item{\code{K}: Number of thresholds for the response categories of the items. If the items are dichotomous \code{K} is \code{NULL}.}
}
}
\description{
Develop a test or a short form given the parameters of dichotomous or polytomous in an item bank/full-length test according to the theta target procedure. See \code{Details}.
}
\details{
Let \eqn{\Theta} be the set of \eqn{N} \eqn{\theta}-targets (\eqn{\theta'}), i.e., the latent trait levels of interest for the assessment defined as discrete levels of \eqn{\theta}, where \eqn{N} denotes the desired length of \eqn{Q_{\text{target}} \subseteq B}, where \eqn{B} denotes the item bank.
The test \eqn{Q_{\text{target}} \subseteq B} of length \eqn{N} is developed by considering the information \eqn{I_i(\theta_n')} that each item in \eqn{B} provides with respect to each \eqn{\theta' \in \Theta}. Therefore, an optimal item (i.e., the item with the highest information function) is chosen for each \eqn{\theta'}. Given that \eqn{N} is the pre-defined length of test \eqn{Q} and that an optimal item is selected for each \eqn{\theta' \in \Theta}, then \eqn{|\Theta| = N}.

Let \eqn{t = 0, \dots, T} denote the iteration index of the procedure and define:
\itemize{
  \item \eqn{Q^t \subseteq B} as the set of items selected for
  inclusion in the test up to iteration \eqn{t};

  \item \eqn{S^t \subseteq \Theta} as the set of  \eqn{\theta'}s
  satisfied up to iteration \eqn{t}.
}

At \eqn{t = 0}, \eqn{Q^0 = \emptyset} and
\eqn{S^0 = \emptyset}.

The procedure iterates the following steps until \eqn{t = T}:

\enumerate{
  \item Select the item--target pair \eqn{(i, n)} maximizing the item
  information function:

  \deqn{
  (i, n) =
  \arg\max_{i \in B \setminus S^t,\; n \in N \setminus Q^t}
  \mathrm{IIF}(i, n)
  }

  \item Update the set of selected items:
  \deqn{
  Q^{t+1} = Q^t \cup \{i\}
  }

  \item Update the set of satisfied ability targets:
  \deqn{
  S^{t+1} = S^t \cup \{n\}
  }
}

At iteration \eqn{T}, the procedure yields
\eqn{|Q^{T+1}| = N} and \eqn{|S^{T+1}| = N}. Further details can be found in Epifania et al. (2022).
}
\examples{
set.seed(123)
n <- 50
theta <- rnorm(500)
item_pars <- data.frame(
  b = runif(n, -3, 3),
  a = runif(n, 1.2, 1.9),
  c = rep(0, n),
  e = rep(1, n)
)
targets <- define_targets(theta, num_targets = 4)
resT <- theta_target(targets, item_pars)
str(resT)
# polytomous items with user defined theta targets
item_pars <- data.frame(matrix(c(
 1.2, 1.0, 0.8,  -1.0, 0.0, 1.2,
 0.9, 1.1, 1.3,  -0.5, 0.7, 1.8,
 0.5, 1.5, 1, -1.5, -1.0, 0,
 1, 1, 1, -1.5, -0, 0.5
 ), nrow = 4, byrow = TRUE))
colnames(item_pars) = paste(rep(c("a", "b"), each = 3), 1:3, sep = "")
resT_poly <- theta_target(c(-1,0), item_pars, K = 3)
str(resT_poly)
}
\references{
Epifania, O. M., Anselmi, P., & Robusto, E. (2022). Item response
theory approaches for test shortening. In M. Wiberg, D. Molenaar,
J. Gonzalez, J. S. Kim, & H. Hwang (Eds.), Quantitative Psychology
(Vol. 422, pp. 75–83). Springer Proceedings in Mathematics and
Statistics. Springer, Cham.
https://doi.org/10.1007/978-3-031-27781-8_7
}
