% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_target.R
\name{theta_target}
\alias{theta_target}
\title{Theta target procedure}
\usage{
theta_target(targets, item_par)
}
\arguments{
\item{targets}{numeric vector with the discrete values of theta for which the information needs to be maximized}

\item{item_par}{dataframe, with nrows equals to the length of the latent trait and four columns, each denoting the IRT item parameters}
}
\value{
A list of length 4 with:

- stf: dataframe with the items selected for inclusion in the STF (column isel), their maximum information function (maxiif), for  a specific theta-target (column theta-target)
- item_par: the original dataframe containing the item parameters
- selected_items: dataframe with the parameters of the selected items
- intervals: reports whether the theta-targets were obtained by clustering the latent trait (clusters) or by dividing it into equal intervals (equal). The label "unknown" identifies any other case.
}
\description{
Procedure based on the theta targets procedure for the generation of a short test form
}
\examples{
set.seed(123)
n <- 50
theta <- rnorm(500)
item_par <- data.frame(
  b = runif(n, -3, 3),
  a = runif(n, 1.2, 1.9),
  c = rep(0, n),
  e = rep(1, n)
)
targets <- define_targets(theta, num_targets = 4)
resT <- theta_target(targets, item_par)
str(resT)
}
