% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_info.R
\name{tif}
\alias{tif}
\title{Test Information Function}
\usage{
tif(iifs, fun = "sum", theta = seq(-5, 5, length.out = 1000))
}
\arguments{
\item{iifs}{matrix or dataframe with nrows equal to the levesl of theta and ncols equal to the number of items in a test}

\item{fun}{character, defines the function for the computation of the TIF, either by summing the items (sum) or by computing the mean (mean)}

\item{theta}{numeric vector (might also be a single value) representing the latent trait. Default is a vector of 1 thousand values rangin from -5 to +5}
}
\value{
A dataframe with two columns: theta with the values of theta, either default values or defined by the user, and the value of the TIF
}
\description{
Computes the test information function of a test given a matrix of IIF
}
\examples{
set.seed(123)
n <- 5
item_par <- data.frame(
  b = runif(n, -3, 3),
  a = runif(n, 1.2, 1.9),
  c = rep(0, n),
  e = rep(1, n)
)
iifs <- item_info(item_par)
test_tif <- tif(iifs)
}
